requirements.txt
# Install inside a Python 3.10 conda env (PROJECT_PLAN_v10.md §4.1).
# After a successful full install, regenerate a lockfile with:
#   pip freeze > requirements.txt
#
# --- Pinned core (§4.3) ---
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
timm==0.9.12
facenet-pytorch==2.5.2
mediapipe==0.10.9
# Linux / GPU server: pip wheel (§4.2). macOS arm64: use `conda install -c conda-forge opencv` instead.
opencv-python==4.9.0.80 ; sys_platform == "linux"
# Linux GPU server only (§4.2) — not expected to install on macOS arm64.
insightface>=0.7 ; sys_platform == "linux"
onnxruntime-gpu>=1.16.0 ; sys_platform == "linux"

# --- From §4.1 (compatible ranges where the plan does not pin) ---
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
wandb>=0.16.0
tensorboard>=2.14.0
PyYAML>=6.0.1
tqdm>=4.66.0
pytorch-grad-cam>=1.4.8
xgboost>=2.0.0
streamlit>=1.28.0
fpdf2>=2.7.0
Pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
flask>=3.0.0
requests>=2.31.0

# --- Dev (§4.1 / NFR-09) ---
pytest>=7.4.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
pre-commit>=3.6.0

# --- V2-alpha FastAPI service (V2A-01) — pinned per docs/AGENTS.md / IMPLEMENTATION_PLAN ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.13.3
redis==5.1.1
rq==1.16.2
fakeredis==2.26.1
httpx==0.27.2
boto3==1.35.36
python-multipart==0.0.12
slowapi==0.1.9
